fix: waiver/FAAB consistency
- get_waiver_targets: message counts claim_now/add_now only (N worth a claim, M speculative); values scaled by league scoring like recommend_faab_bid; ros_gain explained; drop note says 'no market value' instead of 'value 0'
- recommend_faab_bid: upgrade/tier/warnings from the lineup-point horizons when known; weeks_left counts the current unplayed week; non-FAAB leagues return null bid_pct/range_pct
- free agents: claim_processes_at null ('instant'), claim_cost_places 0

+++ nfl_mcp/waiver_target_tools.py
@@ -493,7 +493,7 @@ async def get_waiver_targets(
             hit = service.lookup(values, player_id=p.get("player_id"),
                                  name=p.get("name"), position=p.get("position"))
             mult = model.value_multiplier(p.get("position")) if model is not None else 1.0
-            p["value"] = (int(round(float(hit["value"]) * mult))
+            p["value"] = (round(float(hit["value"]) * mult)
                           if hit and hit.get("value") is not None else None)
         values_ok = bool((values or {}).get("list"))
     except Exception as e:
